# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: NeuroNotes API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /users/profile:
    get:
      operationId: getUserProfile
      tags: [users]
      summary: Get the current user profile
      responses:
        "200":
          description: User profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"
        "404":
          description: User not found

    post:
      operationId: upsertUserProfile
      tags: [users]
      summary: Create or update user profile (onboarding)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpsertUserProfileBody"
      responses:
        "200":
          description: Updated user profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"
...
components:
...
  schemas:
...
    UserProfile:
      type: object
      properties:
        id:
          type: string
        email:
          type: ["string", "null"]
        role:
          type: ["string", "null"]
        goal:
          type: ["string", "null"]
        degree:
          type: ["string", "null"]
        referralSource:
          type: ["string", "null"]
        learnerRole:
          type: ["string", "null"]
        learningGoals:
          type: array
          items:
            type: string
        studyFocus:
          type: ["string", "null"]
        epppInterest:
          type: ["string", "null"]
        selectedTier:
          type: ["string", "null"]
        selectedProduct:
          type: ["string", "null"]
        subscriptionStatus:
          type: string
        onboardingComplete:
          type: boolean
        onboardingCompletedAt:
          type: ["string", "null"]
        stripeCustomerId:
          type: ["string", "null"]
        stripeSubscriptionId:
          type: ["string", "null"]
        createdAt:
          type: string
      required:
        - id
        - subscriptionStatus
        - onboardingComplete
        - learningGoals
    UpsertUserProfileBody:
      type: object
      properties:
        email:
          type: string
        role:
          type: string
        goal:
          type: string
        degree:
          type: string
        referralSource:
          type: string
        learnerRole:
          type: string
        learningGoals:
          type: array
          items:
            type: string
        studyFocus:
          type: string
        epppInterest:
          type: string
        selectedTier:
          type: string
        selectedProduct:
          type: string
        onboardingComplete:
          type: boolean
...